move drawing extensions to a (non-namespace-provider) subfolder
