+++ src/ImageSharp.Drawing/Processing/Extensions/DrawBezierExtensions.cs
