+++ src/ImageSharp.Drawing/Processing/Extensions/DrawPathCollectionExtensions.cs
