+++ src/ImageSharp.Drawing/Processing/Extensions/DrawPathExtensions.cs
