+++ src/ImageSharp.Drawing/Processing/Extensions/DrawPolygonExtensions.cs
