+++ src/ImageSharp.Drawing/Processing/Extensions/DrawRectangleExtensions.cs
