+++ src/ImageSharp.Drawing/Processing/Extensions/FillPathBuilderExtensions.cs
