+++ src/ImageSharp.Drawing/Processing/Extensions/FillPathCollectionExtensions.cs
