+++ src/ImageSharp.Drawing/Processing/Extensions/FillPathExtensions.cs
