+++ src/ImageSharp.Drawing/Processing/Extensions/FillPolygonExtensions.cs
