+++ src/ImageSharp.Drawing/Processing/Extensions/FillRectangleExtensions.cs
