+++ src/ImageSharp.Drawing/Processing/Extensions/FillRegionExtensions.cs
